requirements.txt
fastapi
pydantic
pydantic-settings
asyncpg
redis
loguru
requests
uvicorn
sqlalchemy
alembic
psycopg2-binary
asyncpg
psycopg2
python-jose
environs